@@ -7,7 +7,7 @@ namespace ATMAPI.Services
         void AddAccount(Account account); 
         ICollection<Account> GetAccounts();
         Account GetAccountByNumber(long accountNumber);
-        void UpdateAccount(Account updatedAccount);
+        void UpdateAccount(Account  updatedAccount);
 
 
     }
